requirements.txt
anthropic
requests
python-dotenv
beautifulsoup4
